Updated Volume & Consul Images (non-arm compose)
Modified the compose files to use the new images I was building locally, and
I modified the Consul service specifically with the "command" directive
that I took out of its dockerfile. I appended the flag "-config-dir" with a
value on to the command, and now it will use a secure/non-secure directory
when loading config files instead of a single directory.