Updated compose files
Cleaned up these compose files a little more by taking out the custom
command lines for the consul service because the entrypoint script
was adjusted in the image so that it's now transparent when the image
is loaded. I also added in environment options to the consul services
that will inject non-default values in to the docker image at run time
to determine what db consul should be checking and if it should be
checking for secure services.